implementation/central-api/requirements.txt
# O-Red Central API - Dependencies

# FastAPI et serveur
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de données
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1

# Authentification et sécurité
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# Cryptographie post-quantique
pycryptodome==3.19.0
liboqs-python==0.9.0

# Communication P2P et réseau
aiohttp==3.9.1
websockets==12.0
aiodns==3.1.1

# Validation et sérialisation
pydantic==2.5.0
pydantic-settings==2.1.0

# Cache et performance
redis==5.0.1
aioredis==2.0.1

# Monitoring et métriques
prometheus-client==0.19.0
structlog==23.2.0

# IA et Machine Learning
torch==2.1.1
transformers==4.36.0
sentence-transformers==2.2.2

# Outils de développement
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Production
gunicorn==21.2.0
httpx==0.25.2